Remove duplicate card checkbox component.
